+++ kitty/session.py
@@ -9,13 +9,12 @@ from functools import partial
 from typing import TYPE_CHECKING, Callable, Generator, Iterator, List, Mapping, Optional, Tuple, Union
 
 from .cli_stub import CLIOptions
-from .constants import kitten_exe
 from .layout.interface import all_layouts
 from .options.types import Options
 from .options.utils import resize_window, to_layout_names, window_size
 from .os_window_size import WindowSize, WindowSizeData, WindowSizes
 from .typing import SpecialWindowInstance
-from .utils import expandvars, log_error, resolve_custom_file, resolved_shell, which
+from .utils import cmdline_for_hold, expandvars, log_error, resolve_custom_file, resolved_shell, which
 
 if TYPE_CHECKING:
     from .launch import LaunchSpec
@@ -238,7 +237,7 @@ def create_sessions(
         cmd = args.args if args and args.args else resolved_shell(opts)
         if args and args.hold:
             cmd[0] = which(cmd[0]) or cmd[0]
-            cmd = [kitten_exe(), '__hold_till_enter__'] + cmd
+            cmd = cmdline_for_hold(cmd)
         from kitty.tabs import SpecialWindow
         cwd: Optional[str] = args.directory if respect_cwd and args else None
         special_window = SpecialWindow(cmd, cwd_from=cwd_from, cwd=cwd)